+++ baiducloud-java-sdk-dbsc/src/main/java/com/baidubce/dbsc/sample/ExampleGetMysqlActiveSessions.java
@@ -4,7 +4,6 @@ import com.baidubce.BceClientConfiguration;
 import com.baidubce.BceClientException;
 import com.baidubce.auth.DefaultBceCredentials;
 import com.baidubce.dbsc.DbscClient;
-import com.baidubce.dbsc.models.GetMysqlActiveSessionsRequest;
 import com.baidubce.dbsc.models.GetMysqlActiveSessionsResponse;
 
 public class ExampleGetMysqlActiveSessions {
@@ -19,10 +18,8 @@ public class ExampleGetMysqlActiveSessions {
         bceClientConfig.setCredentials(new DefaultBceCredentials(ak, sk));
 
         DbscClient client = new DbscClient(bceClientConfig);
-        GetMysqlActiveSessionsRequest getMysqlActiveSessionsRequest = new GetMysqlActiveSessionsRequest();
-        getMysqlActiveSessionsRequest.setAppId("");
         try {
-            GetMysqlActiveSessionsResponse response = client.getMysqlActiveSessions(getMysqlActiveSessionsRequest);
+            GetMysqlActiveSessionsResponse response = client.getMysqlActiveSessions();
             System.out.println(response.toJsonString());
         } catch (BceClientException e) {
             System.out.println(e.getMessage());