+++ baiducloud-java-sdk-dbsc/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-09-21 Version: 0.0.2
+- 涉及产品: DBSC，MySQL会话新增appId查询字段
+- 涉及产品: DBSC，Redis慢日志新增clientIp字段
+- 涉及产品: DBSC，覆盖慢日志、限流、空间分析等接口
+
 2026-09-21 Version: 0.0.1
 - 涉及产品: DBSC，新增MySQL慢日志会话限流表空间诊断
 - 涉及产品: DBSC，新增MongoDB慢日志与空间分析